feat: initial commit
initial code in here is from an early
commit from mnbpdx/kickit, wanted to
make something specific with that, but
after following  "Software Engineer Haydn"'s
first video tutorial, I realized I wanted
a general cli utitlity command like his
eve as well. So this is that. it needs
to be renamed, just blanking on that right now.